feat: ILogger class for future database logging
Program.cs:
- add: Configure the DatabaseLogger during startup.

appsettings:
- add: Use the DatabaseLogger for development and in general.

DatabaseLogger:
- add: stub ILogger implementation, for future database logging.
For now, it logs to the console.

The DatabaseLoggerExtensions class is used to create a new
DatabaseLoggerProvider, which creates the custom logger on startup.
When another class needs to create a logger for a specific category,
the provider provides the same logger for each category.

The DatabaseLoggerConfiguration contains an EventId for now, but could
contain more configuration options in the future.