feat/fix: Message class, allocate buffer properly
Notable changes:
- Added convenience Message class for serializing and deserializing
messages between the server and client. The messages can be
(de)serialized from and to JSON.
- The websocket receive() implementation has been placed in it's own
method, and allocates the receive buffer size properly by reading the
size bytes which the client sends upfront.

The way the footage data is send has been changed slightly: before,
each element in the footage array contained a JSON object with only
one key-value pair. This has been changed to just an array of strings,
which represent the filenames. This can be changed in the future to
allow for thumbnails.