Prevent csrf
Update of the most views:
Added request forgery tokens to the form.

In startup the AutoValidateAntiForgeryToken attribute is now passed to each endpoint. This attribute will enable validation of  anti forgery tokens by default on all unsafe requests.

In the AccountIndex.js and CameraIndex.js files I had to add some code to get the tokens from the dom. These tokens will then passed to the request which will delete the account or camera.

In cameracontroller:
Added a title